+++ src/platforms/wordpress/Framework/Gantry.php
@@ -6,6 +6,24 @@ use Gantry\Component\Config\Config;
 class Gantry extends Base\Gantry
 {
     /**
+     * @return boolean
+     */
+    public function debug()
+    {
+        $option = (array) get_option('gantry5_plugin');
+
+        return !empty($option['debug']);
+    }
+
+    /**
+     * @return boolean
+     */
+    public function admin()
+    {
+        return \is_user_logged_in() && \current_user_can('edit_theme_options');
+    }
+
+    /**
      * @param string $location
      * @param bool   $force
      * @return array